Create new ArchiveItem component
